+++ apps/client/src/common/components/input/time-input/TimeInput.module.scss
@@ -1,4 +1,4 @@
 .timeInput {
     letter-spacing: 1px;
-    width: 6.5em;
+    width: 100%;
 }
