+++ apps/client/src/theme/ontimeCheckbox.module.scss
@@ -1,0 +1,5 @@
+.root {
+  .label {
+    color: $gray-500;
+  }
+}
